include/image-commands.mk: extend Build/fit for U-Boot script integration
This commit extends the Build/fit macro so that if a ITS_UBOOT_SCRIPT
variable is defined, it is pass to the mkits.sh script as "-s"
argument. This allows the ITS_UBOOT_SCRIPT file to be integrated as a
U-Boot script in the FIT image.

This will be used as part of the dm-verity integration to add in the
FIT image a U-Boot script that provides the details of the dm-verity
volume (salt, root hash, number of data blocks, start of hash blocks,
etc.).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>