target/linux/generic: add patches to support dm-verity volume at boot
The upstream Linux kernel does not provide a mechanism to setup a DM
volume at boot time using the kernel command line: an initramfs with
user-space tools is strictly required. Since OpenWRT doesn't support
an initramfs as an intermediate step before mounting the real root
filesystem, this requirement of the upstream kernel makes it difficult
to use dm-verity in the context of OpenWRT.

However, some patches have been repeatedly contributed to the upstream
kernel (but not accepted so far) to add a dm="..." kernel argument
that allows to set up a DM device from the kernel command line. This
allows to avoid the need for an initramfs just to set up a DM device
at boot time.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>